+++ RedPandaIDE/settings.cpp
@@ -4711,32 +4711,15 @@ void Settings::CodeCompletion::setShowCodeIns(bool newShowCodeIns)
     mShowCodeIns = newShowCodeIns;
 }
 
-//bool Settings::CodeCompletion::clearWhenEditorHidden()
-//{
-//    if (!mShareParser) {
-//#ifdef Q_OS_WIN
-//        MEMORYSTATUSEX statex;
-//        statex.dwLength = sizeof (statex);
-//        GlobalMemoryStatusEx (&statex);
+bool Settings::CodeCompletion::clearWhenEditorHidden()
+{
+    return mClearWhenEditorHidden;
+}
 
-//        if (statex.ullAvailPhys < (long long int)2*1024*1024*1024) {
-//            return true;
-//        }
-//#elif defined(Q_OS_LINUX)
-//        struct sysinfo si;
-//        sysinfo(&si);
-//        if (si.freeram < (long long int)2*1024*1024*1024) {
-//            return true;
-//        }
-//#endif
-//    }
-//    return mClearWhenEditorHidden;
-//}
-
-//void Settings::CodeCompletion::setClearWhenEditorHidden(bool newClearWhenEditorHidden)
-//{
-//    mClearWhenEditorHidden = newClearWhenEditorHidden;
-//}
+void Settings::CodeCompletion::setClearWhenEditorHidden(bool newClearWhenEditorHidden)
+{
+    mClearWhenEditorHidden = newClearWhenEditorHidden;
+}
 
 int Settings::CodeCompletion::minCharRequired() const
 {
@@ -4902,7 +4885,7 @@ void Settings::CodeCompletion::doSave()
     saveValue("ignore_case",mIgnoreCase);
     saveValue("append_func",mAppendFunc);
     saveValue("show_code_ins",mShowCodeIns);
-    //saveValue("clear_when_editor_hidden",mClearWhenEditorHidden);
+    saveValue("clear_when_editor_hidden",mClearWhenEditorHidden);
     saveValue("min_char_required",mMinCharRequired);
     saveValue("hide_symbols_start_with_two_underline", mHideSymbolsStartsWithTwoUnderLine);
     saveValue("hide_symbols_start_with_underline", mHideSymbolsStartsWithUnderLine);
@@ -4930,35 +4913,8 @@ void Settings::CodeCompletion::doLoad()
     mHideSymbolsStartsWithUnderLine = boolValue("hide_symbols_start_with_underline", true);
 
     bool shouldShare= true;
-//    bool doClear = false;
-
-//#ifdef Q_OS_WIN
-//    MEMORYSTATUSEX statex;
-
-//    statex.dwLength = sizeof (statex);
-
-//    GlobalMemoryStatusEx (&statex);
-
-//    if (statex.ullAvailPhys > (long long int)32*1024*1024*1024) {
-//        shouldShare = false;
-//    }
-
-////    if (shouldShare) {
-////        SYSTEM_INFO info;
-////        GetSystemInfo(&info);
-////        if (info.dwNumberOfProcessors>8 && info.dwProcessorType) {
-////            doClear = true;
-////        }
-////    }
-//#elif defined(Q_OS_LINUX)
-//    struct sysinfo si;
-//    sysinfo(&si);
-//    if (si.freeram > (long long int)24*1024*1024*1024) {
-//        shouldShare = false;
-//    }
-//#endif
-    //mClearWhenEditorHidden = boolValue("clear_when_editor_hidden",doClear);
     mShareParser = boolValue("share_parser",shouldShare);
+    mClearWhenEditorHidden = boolValue("clear_when_editor_hidden", mShareParser);
 }
 
 Settings::CodeFormatter::CodeFormatter(Settings *settings):