+++ RedPandaIDE/settingsdialog/environmentperformancewidget.cpp
@@ -33,32 +33,13 @@ EnvironmentPerformanceWidget::~EnvironmentPerformanceWidget()
 
 void EnvironmentPerformanceWidget::doLoad()
 {
-//    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
-//#ifdef Q_OS_WIN
-//    MEMORYSTATUSEX statex;
-
-//    statex.dwLength = sizeof (statex);
-
-//    GlobalMemoryStatusEx (&statex);
-//    if (statex.ullTotalPhys < (long long int)2*1024*1024*1024) {
-//        ui->chkClearWhenEditorHidden->setEnabled(false);
-//        ui->chkClearWhenEditorHidden->setChecked(true);
-//        pSettings->codeCompletion().setClearWhenEditorHidden(true);
-//        pSettings->codeCompletion().save();
-//    }
-//    if (statex.ullTotalPhys < (long long int)1024*1024*1024) {
-//        ui->chkEditorsShareParser->setEnabled(false);
-//        ui->chkEditorsShareParser->setChecked(true);
-//        pSettings->codeCompletion().setShareParser(true);
-//        pSettings->codeCompletion().save();
-//    }
-//#endif
+    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
     ui->chkEditorsShareParser->setChecked(pSettings->codeCompletion().shareParser());
 }
 
 void EnvironmentPerformanceWidget::doSave()
 {
-    //pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
+    pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
     pSettings->codeCompletion().setShareParser(ui->chkEditorsShareParser->isChecked());
 
     pSettings->codeCompletion().save();